Converted forms. As-is.
