+++ src/Asm.MooBank.Web.App/package.json
@@ -14,7 +14,7 @@
   },
   "type": "module",
   "dependencies": {
-    "@andrewmclachlan/mooapp": "^2.2.800",
+    "@andrewmclachlan/mooapp": "^2.2.801",
     "@andrewmclachlan/mooicons": "^2.2.774",
     "@fortawesome/fontawesome-svg-core": "^6.7.2",
     "@fortawesome/free-solid-svg-icons": "^6.7.2",
