+++ src/Asm.MooBank.Web.App/src/pages/accounts/index.ts
@@ -1,3 +1,5 @@
 export * from "./Account";
 export * from "./Accounts";
+export * from "./CreateAccount";
+export * from "./ManageAccount";
 export * from "./VirtualAccount";
