+++ src/Asm.MooBank.Web.App/src/pages/assets/ManageAsset.tsx
@@ -1,79 +1,73 @@
 import React, { useEffect, useState } from "react";
-import { Button, Form, InputGroup } from "react-bootstrap";
+import { Button, InputGroup } from "react-bootstrap";
 import { useNavigate } from "react-router";
+
+import { Form, SectionForm } from "@andrewmclachlan/mooapp";
+
 import { Asset } from "../../models";
 import { AssetPage } from "./AssetPage";
 import { useAsset } from "./AssetProvider";
 
-import { useGroups, useUpdateAsset } from "services";
+import { useUpdateAsset } from "services";
+import { GroupSelector } from "components/GroupSelector";
+import { useForm } from "react-hook-form";
 
 export const ManageAsset: React.FC = () => {
 
     const navigate = useNavigate();
 
-    const { data: groups } = useGroups();
     const updateAsset = useUpdateAsset();
 
-    const currentAsset = useAsset();
+    const asset = useAsset();
 
-    useEffect(() => {
-        setAsset(currentAsset);
-    }, [currentAsset]);
+    const handleSubmit = (data: Asset) => {
 
-    const [asset, setAsset] = useState<Asset>();
-    const handleSubmit = (e: React.FormEvent<HTMLFormElement>) => {
-        e.stopPropagation();
-        e.preventDefault();
-
-        updateAsset(asset);
+        updateAsset(data);
 
         navigate("/accounts");
     }
 
-    if (!asset) return null;
+    const form = useForm({ defaultValues: asset });
+
+    useEffect(() => {
+        form.reset(asset);
+    }, [asset, form]);
 
     return (
-        <AssetPage title="Manage" breadcrumbs={[{ text: "Manage", route: `/shares/${asset.id}/manage` }]}>
-            <Form className="section" onSubmit={handleSubmit}>
-                <Form.Group controlId="accountName" >
+        <AssetPage title="Manage" breadcrumbs={[{ text: "Manage", route: `/shares/${asset?.id}/manage` }]}>
+            <SectionForm form={form} onSubmit={handleSubmit}>
+                <Form.Group groupId="name" >
                     <Form.Label>Name</Form.Label>
-                    <Form.Control type="text" required maxLength={50} value={asset.name} onChange={(e: any) => setAsset({ ...asset, name: e.currentTarget.value })} />
-                    <Form.Control.Feedback type="invalid">Please enter a name</Form.Control.Feedback>
+                    <Form.Input required maxLength={50} />
                 </Form.Group>
-                <Form.Group controlId="accountDescription" >
+                <Form.Group groupId="description" >
                     <Form.Label>Description</Form.Label>
-                    <Form.Control type="text" as="textarea" required maxLength={255} value={asset.description} onChange={(e: any) => setAsset({ ...asset, description: e.currentTarget.value })} />
-                    <Form.Control.Feedback type="invalid">Please enter a description</Form.Control.Feedback>
+                    <Form.TextArea required maxLength={255} />
                 </Form.Group>
-                <Form.Group controlId="price" >
+                <Form.Group groupId="purchasePrice" >
                     <Form.Label>Purchase Price</Form.Label>
                     <InputGroup>
                         <InputGroup.Text>$</InputGroup.Text>
-                        <Form.Control type="number" required value={asset.purchasePrice?.toString()} onChange={(e: any) => setAsset({ ...asset, purchasePrice: e.currentTarget.value })} />
+                        <Form.Input type="number" required />
                     </InputGroup>
                 </Form.Group>
-                <Form.Group controlId="value" >
+                <Form.Group groupId="currentBalance" >
                     <Form.Label>Current Value</Form.Label>
                     <InputGroup>
                         <InputGroup.Text>$</InputGroup.Text>
-                        <Form.Control type="number" required value={asset.currentBalance.toString()} onChange={(e: any) => setAsset({ ...asset, currentBalance: e.currentTarget.value})} />
+                        <Form.Input type="number" required />
                     </InputGroup>
                 </Form.Group>
-                <Form.Group controlId="group">
+                <Form.Group groupId="groupId">
                     <Form.Label>Group</Form.Label>
-                    <Form.Select value={asset.groupId} onChange={(e: any) => setAsset({ ...asset, groupId: e.currentTarget.value })}>
-                        <option value="">Select a group...</option>
-                        {groups?.map(a =>
-                            <option value={a.id} key={a.id}>{a.name}</option>
-                        )}
-                    </Form.Select>
+                    <GroupSelector />
                 </Form.Group>
-                <Form.Group controlId="ShareWithFamily">
+                <Form.Group groupId="shareWithFamily">
                     <Form.Label>Visible to other family members</Form.Label>
-                    <Form.Check checked={asset.shareWithFamily} onChange={(e: any) => setAsset({ ...asset, shareWithFamily: e.currentTarget.checked })} />
+                    <Form.Check />
                 </Form.Group>
                 <Button type="submit" variant="primary">Update</Button>
-            </Form>
+            </SectionForm>
         </AssetPage>
     );
 }
