--- src/Asm.MooBank.Web.App/src/pages/createAccount/index.ts
@@ -1,1 +1,0 @@
-export * from "./CreateAccount";
