+++ src/Asm.MooBank.Web.App/src/pages/index.ts
@@ -2,12 +2,10 @@ export * from "./accounts";
 export * from "./assets";
 export * from "./bills";
 export * from "./budget";
-export * from "./createAccount";
 export * from "./dashboard";
 export * from "./error";
 export * from "./groups";
 export * from "./Import";
-export * from "./manageAccounts";
 export * from "./reports";
 export * from "./rules/Rules";
 export * from "./settings";
