--- src/Asm.MooBank.Web.App/src/pages/manageAccounts/index.ts
@@ -1,2 +1,0 @@
-export * from "./ManageAccount";
-export * from "./ManageAccounts";
