+++ src/Asm.MooBank.Web.App/src/pages/stockHoldings/CreateStockHolding.tsx
@@ -1,44 +1,26 @@
 import React, { useState } from "react";
-import { Button, Form, InputGroup } from "react-bootstrap";
+import { Button, InputGroup } from "react-bootstrap";
 import { useNavigate } from "react-router";
 import { NewStockHolding, emptyStockHolding } from "../../models";
 
-import { Page } from "@andrewmclachlan/mooapp";
+import { Form, Page, SectionForm } from "@andrewmclachlan/mooapp";
 import { useCreateStockHolding, useGroups } from "services";
+import { useForm } from "react-hook-form";
+import { GroupSelector } from "components/GroupSelector";
 
 export const CreateStockHolding: React.FC = () => {
 
     const navigate = useNavigate();
 
-    const { data: groups } = useGroups();
     const createStockHolding = useCreateStockHolding();
 
-    const [name, setName] = useState("");
-    const [description, setDescription] = useState("");
-    const [symbol, setSymbol] = useState("");
-    const [quantity, setQuantity] = useState(0);
-    const [price, setPrice] = useState(0);
-    const [fees, setFees] = useState(0);
-    const [groupId, setGroupId] = useState("");
-    const [shareWithFamily, setShareWithFamily] = useState(false);
 
-    const handleSubmit = (e: React.FormEvent<HTMLFormElement>) => {
-        e.stopPropagation();
-        e.preventDefault();
+    const handleSubmit = (data: NewStockHolding) => {
 
         const stockHolding: NewStockHolding = {
-            ...emptyStockHolding,
-            name: name,
-            symbol: symbol,
-            quantity: quantity,
-            price: price,
-            fees: fees,
-            description: description,
+            ...data,
             controller: "Manual",
             currency: "AUD",
-            currentBalanceLocalCurrency: 0,
-            groupId: groupId === "" ? undefined : groupId,
-            shareWithFamily: shareWithFamily,
             currentBalance: 0,
         };
 
@@ -47,59 +29,53 @@ export const CreateStockHolding: React.FC = () => {
         navigate("/accounts");
     }
 
+    const form = useForm({defaultValues: emptyStockHolding});
+
     return (
         <Page title="Create Shares" breadcrumbs={[{ text: "Accounts", route: "/accounts" }, { text: "Create Shares", route: "/stock/create" }]}>
-            <Form className="section" onSubmit={handleSubmit}>
-                <Form.Group controlId="accountName" >
+            <SectionForm form={form} onSubmit={handleSubmit}>
+                <Form.Group groupId="name" >
                     <Form.Label>Name</Form.Label>
-                    <Form.Control type="text" required maxLength={50} value={name} onChange={(e: any) => setName(e.currentTarget.value)} />
-                    <Form.Control.Feedback type="invalid">Please enter a name</Form.Control.Feedback>
+                    <Form.Input required maxLength={50} />
                 </Form.Group>
-                <Form.Group controlId="accountDescription" >
+                <Form.Group groupId="description" >
                     <Form.Label>Description</Form.Label>
-                    <Form.Control type="text" as="textarea" required maxLength={255} value={description} onChange={(e: any) => setDescription(e.currentTarget.value)} />
-                    <Form.Control.Feedback type="invalid">Please enter a description</Form.Control.Feedback>
+                    <Form.TextArea required maxLength={255} />
                 </Form.Group>
-                <Form.Group controlId="symbol" >
+                <Form.Group groupId="symbol" >
                     <Form.Label>Symbol</Form.Label>
-                    <Form.Control type="text" required value={symbol} onChange={(e: any) => setSymbol(e.currentTarget.value)} />
-                    <Form.Control.Feedback type="invalid">Please enter a symbol</Form.Control.Feedback>
+                    <Form.Input required />
                 </Form.Group>
-                <Form.Group controlId="quantity" >
+                <Form.Group groupId="quantity" >
                     <Form.Label>Opening Quantity</Form.Label>
                     <InputGroup>
-                        <Form.Control type="number" required value={quantity.toString()} onChange={(e: any) => setQuantity(e.currentTarget.value)} />
+                        <Form.Input type="number" required />
                     </InputGroup>
                 </Form.Group>
-                <Form.Group controlId="price" >
+                <Form.Group groupId="price" >
                     <Form.Label>Purchase Price</Form.Label>
                     <InputGroup>
                         <InputGroup.Text>$</InputGroup.Text>
-                        <Form.Control type="number" required value={price.toString()} onChange={(e: any) => setPrice(e.currentTarget.value)} />
+                        <Form.Input type="number" required />
                     </InputGroup>
                 </Form.Group>
-                <Form.Group controlId="quantity" >
+                <Form.Group groupId="quantity" >
                     <Form.Label>Brokerage Fees</Form.Label>
                     <InputGroup>
                         <InputGroup.Text>$</InputGroup.Text>
-                        <Form.Control type="number" required value={fees.toString()} onChange={(e: any) => setFees(e.currentTarget.value)} />
+                        <Form.Input type="number" required />
                     </InputGroup>
                 </Form.Group>
-                <Form.Group controlId="group">
+                <Form.Group groupId="groupId">
                     <Form.Label>Group</Form.Label>
-                    <Form.Select value={groupId} onChange={(e: any) => setGroupId(e.currentTarget.value)}>
-                        <option value="">Select a group...</option>
-                        {groups?.map(a =>
-                            <option value={a.id} key={a.id}>{a.name}</option>
-                        )}
-                    </Form.Select>
+                    <GroupSelector />
                 </Form.Group>
-                <Form.Group controlId="ShareWithFamily">
+                <Form.Group groupId="ShareWithFamily">
                     <Form.Label>Visible to other family members</Form.Label>
-                    <Form.Check checked={shareWithFamily} onChange={(e: any) => setShareWithFamily(e.currentTarget.checked)} />
+                    <Form.Check />
                 </Form.Group>
                 <Button type="submit" variant="primary">Create</Button>
-            </Form>
+            </SectionForm>
         </Page>
     );
 }
